Added registration page
+database.py
Ability to add, check and login with it. Passwords are stored in hashed hex values in the special column

+registration.ui
Quite responsive ui that will be further modified. ~70% done for now

-customs.py
Deleted all useless exceptions

+Accounts.db
sqlite database where all the users are stored

+registrationPage.py
script to open, render the page and registration logic.